Fix "null" page on /meetups/[slug] page
Some meetup pages displayed "null" because of the
excerpt field not being handled.

In this commit this is fixed by using a getOrElse()
function that will return a fallback value (an empty string
be default) if the value passed is not defined.

In addition to this the description is added to the view as a
paragraph.